+++ prisma/migrations/20240423172533_add_phone_numbers_to_user_root/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE `USERS_ROOT` ADD COLUMN `bot_phone` VARCHAR(20) NULL,
+    ADD COLUMN `human_phone` VARCHAR(20) NULL;