+++ src/components/transactions/services.ts
@@ -221,6 +221,9 @@ export class FinanceServices {
     from?: Date,
     page = 1,
   ): Promise<AlqMovementResponse | undefined> {
+    // TODO
+    // Look up by clave_rastreo with
+    // searchParams.set("clave_rastreo", tracking_number)
     const accountId = CONFIG.EXTERNAL.ALQ_SAVINGS_ACCOUNT_ID;
     const endpoint = `cuenta-ahorro-cliente/${accountId}/transaccion`;
     const PAGE_SIZE = 20;