+++ src/routes/index.ts
@@ -6,6 +6,7 @@ import transactionsRouter from "./transactions.router";
 import agentRouter from "./agent.router";
 import authRouter from "./auth.router";
 import webPushRouter from "./web-push";
+import botRouter from "./bot.router";
 
 const mainRouter = Router();
 
@@ -16,5 +17,6 @@ mainRouter.use("/transactions", transactionsRouter);
 mainRouter.use("/agent", agentRouter);
 mainRouter.use("/auth", authRouter);
 mainRouter.use("/web-push", webPushRouter);
+mainRouter.use("/bot", botRouter);
 
 export default mainRouter;