+++ CHANGELOG.md
@@ -1,7 +1,17 @@
+## [1.0.0] - 2024-04-24
+
+### Changed
+- Moved `/agent/qr/:name` to `/bot/:name?`.
+- Return bot names from `/bot` when `:name` param is omitted.
+
 ## [1.0.0] - 2024-04-23
+
+### Added
+- GET and POST `/agent/support` to view and update support telephone numbers. 
 
 ### Changed
 - Improve `seed.ts` to allow updating credentials
+- Move bot seeding logic into bot repo
 
 ## [1.0.0] - 2024-04-20